Version 1.0

+++ backend/models/portfolio.go
@@ -2,6 +2,6 @@ package models
 
 type PortfolioItem struct {
 	Symbol string  `json:"symbol"`
-	Lots   int     `json:"lots"`
+	Lots   float64 `json:"lots"`
 	Price  float64 `json:"price"`
 }
